remove six module
